% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdata.R
\name{ADToRatio}
\alias{ADToRatio}
\title{Return heteroplasmy from value of AD.}
\usage{
ADToRatio(strAD)
}
\arguments{
\item{strAD}{Value of AD as string.}
}
\value{
Heteroplasmy as double.
}
\description{
Return heteroplasmy from value of AD.
}
