% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdata.R
\name{ExtractFromINFO}
\alias{ExtractFromINFO}
\title{Extract the specified element from INFO section.}
\usage{
ExtractFromINFO(targetTibble, query, colName = query)
}
\arguments{
\item{targetTibble}{Target table as tibble.}

\item{query}{String. the element that you want to extract.}

\item{colName}{Name of column that is being added.}
}
\description{
Extract the specified element from INFO section.
}
