% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdata.R
\name{ExtractINFO}
\alias{ExtractINFO}
\title{Extract each values from INFO section.}
\usage{
ExtractINFO(target_table)
}
\arguments{
\item{target_table}{table from VCF file.}
}
\description{
Extract each values from INFO section.
}
