% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitdata.R
\name{VCFToTibble}
\alias{VCFToTibble}
\title{read VCF file as tibble}
\usage{
VCFToTibble(VCFFile)
}
\arguments{
\item{VCFFile}{VCF file path}
}
\value{
Tibble object
}
\description{
read VCF file as tibble
}
